running_agents/pyproject.toml
[project]
name = "basic-cover"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.2.3",
]
